+++ ziface/imessage.go
@@ -17,13 +17,13 @@ package ziface
 	将请求的一个消息封装到message中，定义抽象层接口
 */
 type IMessage interface {
-	GetDataLen() uint32 //获取消息数据段长度
-	GetMsgID() uint32   //获取消息ID
+	GetDataLen() uint16 //获取消息数据段长度
+	GetMsgID() uint16   //获取消息ID
 	GetData() []byte    //获取消息内容
 
-	SetMsgID(uint32)   //设计消息ID
+	SetMsgID(uint16)   //设计消息ID
 	SetData([]byte)    //设计消息内容
 	SetMsgType(int)    //设置消息类型 websocket
-	SetDataLen(uint32) //设置消息数据段长度
+	SetDataLen(uint16) //设置消息数据段长度
 	ToString()         //输出日志
 }
