+++ znet/request.go
@@ -19,6 +19,6 @@ func (r *Request) GetData() []byte {
 }
 
 //GetMsgID 获取请求的消息的ID
-func (r *Request) GetMsgID() uint32 {
+func (r *Request) GetMsgID() uint16 {
 	return r.msg.GetMsgID()
 }
